+++ csrc/scheduler/pointwise_utils.cpp
@@ -283,7 +283,9 @@ BreakPointInfo getBreakPoint(
       // If outer broadcast, or balanced broadcast:
       if (lhs_bit_multiple <= rhs_bit_multiple &&
           // If right transfer size is bigger than half of L2
-          at::cuda::getCurrentDeviceProperties()->l2CacheSize * 8 <
+          static_cast<int64_t>(
+              at::cuda::getCurrentDeviceProperties()->l2CacheSize) *
+                  8 <
               right_transfer_size_bit * 2) {
         // flip BIDx and BIDy bindings
         result.flip_grid_binding = true;
